requirements.txt
# ── Face Recognition ──────────────────────────────────────────
insightface==0.7.3
onnxruntime==1.16.3

# ── Computer Vision ───────────────────────────────────────────
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# ── Deep Learning (Spoof Detection) ───────────────────────────
torch==2.0.1
torchvision==0.15.2

# ── Image Processing ──────────────────────────────────────────
Pillow==9.3.0
scikit-image==0.21.0
numpy==1.24.4

# ── Database ──────────────────────────────────────────────────
mysql-connector-python==8.2.0

# ── Utilities ─────────────────────────────────────────────────
protobuf==3.20.1
reportlab==4.0.8
python-dotenv==1.0.0